fix check_draining_servers - works

+++ app/jobs/autoscaling.py
@@ -62,7 +62,7 @@ def retry(operation, randgen=random.Random().random):
 
 @retry
 def aws_command(command_str):
-    return subprocess.check_output(command_str.split())
+    return subprocess.check_output(command_str, shell=True)
 
 
 def find_asg(asg_list, name_prefix):
@@ -267,7 +267,7 @@ def get_draining_servers(asg):
     ''' returns a map of draining instance IDs to the time they started draining '''
     cmd = "aws ec2 describe-tags --filters 'Name=tag-key,Values={draining_tag}' 'Name=resource-id,Values={instance_ids}'"
     cmd = cmd.format(draining_tag=DRAINING_TAG, instance_ids=','.join(instances_in(asg)))
-    filtered_tag_dict = json.loads(aws_command(cmd))
+    tag_dict = json.loads(aws_command(cmd))
     instance_dict = { item['ResourceId']: dateutil.parser.parse(item['Value']) for item in tag_dict['Tags'] if item['Key'] == DRAINING_TAG }
     return instance_dict
 
@@ -289,11 +289,15 @@ def count_running_alignment_jobs(asg):
                 count_dict[instance_id] += 1
             else:
                 expired_jobs.append(job_tag)
-    delete_tags_from_instances(instance_ids, expired_jobs)
+    if expired_jobs:
+        print "The following job tags have expired and will be deleted: " + ", ".join(expired_jobs)
+        delete_tags_from_instances(instance_ids, expired_jobs)
     return count_dict
 
 
 def delete_tags_from_instances(instance_ids, tag_keys):
+    if not instance_ids or not tag_keys:
+        return
     cmd = "aws ec2 delete-tags --resources {list_instances} --tags {list_tags}"
     cmd = cmd.format(list_instances=' '.join(instance_ids), list_tags=' '.join(['Key='+tag for tag in tag_keys]))
     aws_command(cmd)
@@ -319,15 +323,16 @@ def check_draining_servers(asg, can_scale):
     drain_date_by_instance_id = get_draining_servers(asg)
     num_jobs_by_instance_id = count_running_alignment_jobs(asg)
     instance_ids_to_terminate = []
-    current_timestamp = iso_time_now()
-    for id, drain_date in drain_date_by_instance_id:
-        draining_interval = current_timestamp - drain_date
+    current_timestamp = dateutil.parser.parse(iso_time_now())
+    for id, drain_date in drain_date_by_instance_id.items():
+        draining_interval = (current_timestamp - drain_date).total_seconds()
         num_jobs = num_jobs_by_instance_id[id]
         if num_jobs == 0 and draining_interval > MAX_REFRESH_INTERVAL:
             instance_ids_to_terminate.append(id)
-    print "Instances {list_instances} are ready to be terminated".format(list_instances=' '.join(instance_ids))
-    if can_scale:
-        remove_termination_protection(instance_ids_to_terminate, asg)
+    if instance_ids_to_terminate:
+        print "The following instances are ready to be terminated: " + ", ".join(instance_ids_to_terminate)
+        if can_scale:
+            remove_termination_protection(instance_ids_to_terminate, asg)
 
 
 def count_running_batch_jobs():
